feat: update select implementation
