--- src/ui/core/select/index.tsx
@@ -1,2 +1,0 @@
-export * from './controlled-select';
-export * from './select';
