--- src/ui/core/select/options.tsx
@@ -1,81 +1,0 @@
-import type { BottomSheetModal } from '@gorhom/bottom-sheet';
-import { BottomSheetFlatList } from '@gorhom/bottom-sheet';
-import { useColorScheme } from 'nativewind';
-import * as React from 'react';
-import { Pressable, type PressableProps } from 'react-native';
-
-import { colors } from '@/ui/theme';
-
-import { Modal } from '../modal';
-import { Text } from '../text';
-import { Check } from './icons';
-
-export type Option = { label: string; value: string | number };
-
-type OptionsProps = {
-  options: Option[];
-  onSelect: (option: Option) => void;
-  value?: string | number;
-};
-
-function keyExtractor(item: Option) {
-  return `select-item-${item.value}`;
-}
-
-export const Options = React.forwardRef<BottomSheetModal, OptionsProps>(
-  ({ options, onSelect, value }, ref) => {
-    const height = options.length * 70 + 100;
-    const snapPoints = React.useMemo(() => [height], [height]);
-    const { colorScheme } = useColorScheme();
-    const isDark = colorScheme === 'dark';
-    const renderSelectItem = React.useCallback(
-      ({ item }: { item: Option }) => (
-        <Option
-          key={`select-item-${item.value}`}
-          label={item.label}
-          selected={value === item.value}
-          onPress={() => onSelect(item)}
-        />
-      ),
-      [onSelect, value]
-    );
-
-    return (
-      <Modal
-        ref={ref}
-        index={0}
-        snapPoints={snapPoints}
-        backgroundStyle={{
-          backgroundColor: isDark ? colors.charcoal[800] : colors.white,
-        }}
-      >
-        <BottomSheetFlatList
-          data={options}
-          keyExtractor={keyExtractor}
-          renderItem={renderSelectItem}
-        />
-      </Modal>
-    );
-  }
-);
-
-const Option = React.memo(
-  ({
-    label,
-    selected = false,
-    ...props
-  }: PressableProps & {
-    selected?: boolean;
-    label: string;
-  }) => {
-    return (
-      <Pressable
-        className="flex-row items-center border-b-[1px] border-neutral-300 bg-white px-3 py-2 dark:border-charcoal-700 dark:bg-charcoal-800"
-        {...props}
-      >
-        <Text className="flex-1 dark:text-charcoal-100 ">{label}</Text>
-        {selected && <Check fill="#000" />}
-      </Pressable>
-    );
-  }
-);
